02_LangChain_Olama.py
from langchain_openai import ChatOpenAI
from langchain_core.prompts import ChatPromptTemplate
from langchain_core.output_parsers import StrOutputParser
from langchain_community.chat_models import ChatOllama

import streamlit as st
from dotenv import load_dotenv

load_dotenv()


prompt = ChatPromptTemplate.from_messages(
    [
        ("system", "You are a helpful assistant. Please respond to user queries."),
        ("user", "Question: {question}")
    ]
)

st.title('Langchain demo with OpenAI API')
input_text = st.text_input("Search the topic you want")


# Ollama Lama2
llm = ChatOllama(model="gemma3:12b")
output_parser = StrOutputParser()   
chain = prompt | llm | output_parser

if input_text:
    response = chain.invoke({"question": input_text})
    st.write(response)