requirements.txt
streamlit 
python-dotenv
langchain==0.2.3 
langchain-core==0.2.5 
langchain-openai==0.1.8 
pydantic==2.7.4
langchain-community==0.2.4